requirements.txt
# Core dependencies
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.0
scipy>=1.7.0
Pillow>=8.0.0
tqdm>=4.62.0
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.3.0
scikit-learn>=1.0.0

# XAI methods
captum>=0.5.0          # For GradCAM, IG, LRP
pytorch-grad-cam>=1.4.0 # For GradCAM visualization
lime>=0.2.0.1          # For LIME
shap>=0.41.0           # For SHAP

# Robustness
robustbench>=1.0       # For robust models
imagecorruptions>=1.1  # For image corruptions

# Additional utilities
pyyaml>=6.0
requests>=2.26.0
psutil>=5.8.0

# Optional but recommended
ipython>=7.30.0
jupyter>=1.0.0
tensorboard>=2.7.0